feat(10-02): extract transfer-funds-core.ts and transfer-token-core.ts
- Create transfer-funds-core.ts with transferFundsCore function and organizationId bypass
- Create transfer-token-core.ts with transferTokenCore and parseTokenAddress functions
- Thin wrappers in step files delegate to core: enrichment + metrics + logging only
- Core files use "server-only" (no "use step"), safe for direct execution API import
